Move python files into their own subdir with a root-level wrapper
