+++ run.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+
+import runpy
+
+runpy.run_module("python.main", run_name="__main__")
